revert package/crate names back
