+++ yew-functional-macro/tests/function_attr_test.rs
