+++ yew-functional-macro/tests/function_macro/applied-to-non-fn-fail.stderr
