+++ yew-functional-macro/tests/function_macro/bad-name-fail.rs
@@ -1,5 +1,5 @@
 use yew::prelude::*;
-use yew_function_components::function_component;
+use yew_functional::function_component;
 
 #[derive(Clone, Properties, PartialEq)]
 struct Props {
