+++ yew-functional-macro/tests/function_macro/bad-name-fail.stderr
