+++ yew-functional-macro/tests/function_macro/bad-props-param-fail.rs
@@ -1,13 +1,13 @@
 use yew::prelude::*;
-use yew_function_components::function_component;
+use yew_functional::function_component;
 
 #[derive(Clone, Properties, PartialEq)]
 struct Props {
     a: usize,
 }
 
-#[function_component()]
-fn comp(props: &Props) -> Html {
+#[function_component(Comp)]
+fn comp(props: Props) -> Html {
     html! {
         <p>
             { props.a }
