+++ yew-functional-macro/tests/function_macro/bad-return-type-fail.stderr
