+++ yew-functional-macro/tests/function_macro/const-fail.stderr
