+++ yew-functional-macro/tests/function_macro/extern-fail.stderr
