+++ yew-functional-macro/tests/function_macro/generic-fail.rs
@@ -1,5 +1,5 @@
 use yew::prelude::*;
-use yew_function_components::function_component;
+use yew_functional::function_component;
 
 #[derive(Clone, Properties, PartialEq)]
 struct Props {
@@ -7,7 +7,7 @@ struct Props {
 }
 
 #[function_component(Comp)]
-fn comp(props: &'static Props) -> Html {
+const fn comp<P: Properties>(props: &P) -> Html {
     html! {
         <p>
             { props.a }
