+++ yew-functional-macro/tests/function_macro/generic-fail.stderr
