+++ yew-functional-macro/tests/function_macro/lifetime-props-param-fail.rs
@@ -1,5 +1,5 @@
 use yew::prelude::*;
-use yew_function_components::function_component;
+use yew_functional::function_component;
 
 #[derive(Clone, Properties, PartialEq)]
 struct Props {
@@ -7,7 +7,7 @@ struct Props {
 }
 
 #[function_component(Comp)]
-fn comp(props: Props) -> Html {
+fn comp(props: &'static Props) -> Html {
     html! {
         <p>
             { props.a }
