+++ yew-functional-macro/tests/function_macro/no-name-fail.rs
@@ -1,13 +1,13 @@
 use yew::prelude::*;
-use yew_function_components::function_component;
+use yew_functional::function_component;
 
 #[derive(Clone, Properties, PartialEq)]
 struct Props {
     a: usize,
 }
 
-#[function_component(Comp)]
-const fn comp<P: Properties>(props: &P) -> Html {
+#[function_component()]
+fn comp(props: &Props) -> Html {
     html! {
         <p>
             { props.a }
