+++ yew-functional-macro/tests/function_macro/no-name-fail.stderr
