+++ yew-functional-macro/tests/function_macro/with-receiver-fail.stderr
