+++ yew-functional-macro/tests/function_macro/without-props-pass.rs
