+++ yew-functional/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "yew-function-components"
+name = "yew-functional"
 version = "0.13.0"
 authors = ["ZainlessBrombie <zainredirect@gmail.com>"]
 edition = "2018"
@@ -17,5 +17,5 @@ yew = { path = "../yew" }
 
 [dependencies]
 yew = { path = "../yew" }
-yew-function-components-macro = { path = "../yew-function-components-macro" }
+yew-functional-macro = { path = "../yew-functional-macro" }
 wasm-bindgen = "0.2.60"
